refactored updateuserProfile
