+++ api/index.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 // Export routes from here
 
-const authRouter = require('./auth/userAuth');
+const authRouter = require('./auth/auth');
 const userRouter = require('./routes/userRoute');
 
 router.use('/auth', authRouter);
