+++ api/middleware/githubStrategy.js
@@ -1,6 +1,6 @@
 const passport = require('passport');
 const GitHubStrategy = require('passport-github').Strategy;
-const User = require('../../database/models/users');
+const model = require('../../database/models');
 const keys = require('../../config/secret');
 
 export default passport.use(
@@ -11,7 +11,7 @@ export default passport.use(
       callbackURL: '/login/?provider=github/redirect'
     },
     (accessToken, refreshToken, profile, cb) => {
-      User.findOrCreate({ githubId: profile.id }, (err, user) => {
+      model.Users.findOrCreate({ Auth_Id: profile.id }, (err, user) => {
         return cb(err, user);
       });
     }
