+++ api/routes/userRoute.js
@@ -1,6 +1,7 @@
 const router = require('express').Router();
 const controller = require('../controllers/user');
 
+router.get('/users', controller.getAllUsers);
 router.post('/:username/profile', controller.createUserProfile);
 router.put('/:username/profile', controller.updateUserProfile);
